+++ database/migrations/2021_01_20_201808_category_product_table.php
@@ -16,8 +16,8 @@ class CategoryProductTable extends Migration
         Schema::create('category_product', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
 
             $table->foreign('category_id')->references('id')->on('categories');
             $table->foreign('product_id')->references('id')->on('products');
